poc1_manual_instrumentation/3rd_part/requirements.txt

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions

requests
opentelemetry-instrumentation-requests
